+++ seedimg-webp/seedimg-webp.hpp
@@ -9,7 +9,7 @@ namespace seedimg::modules {
 namespace webp {
 bool check(const std::string &filename) noexcept;
 bool to(const std::string &filename, std::unique_ptr<seedimg::img> &inp_img,
-        uint8_t quality = 100) noexcept;
+        float quality = 100.0) noexcept;
 std::optional<std::unique_ptr<seedimg::img>>
 from(const std::string &filename) noexcept;
 } // namespace webp
